checkpoint: introduce git access enum
Introduce `GitAccess` enum and attempt to determine whether the user can
actually run `git status` or not in order to display more accurate
information in `GitPanel`.

This doens't seem to be working because we fail to even create the
`Repository` when the user does not own the `.git` folder, so the
scheduled update is never run, hence the field is never updated.

Co-authored-by: cameron <cameron.studdstreet@gmail.com>